move DriverContext to context folder
