+++ frontend/src/pages/Dashboard/Home.tsx
@@ -4,7 +4,7 @@ import UnscheduledTable from '../../components/UserTables/UnscheduledTable';
 import Schedule from '../../components/Schedule/Schedule';
 import MiniCal from '../../components/MiniCal/MiniCal';
 import styles from './page.module.css';
-import {useDrivers, DriversProvider} from '../../components/DriverCards/DriversContext';
+import {useDrivers, DriversProvider} from '../../context/DriversContext';
 import ExportButton from '../../components/ExportButton/ExportButton';
 
 const Drivers = () => {
